prompt_engineering/pyproject.toml
[tool.poetry]
name = "explore-open-ai"
version = "0.1.0"
description = "Project for exploring open ai."
authors = ["Dilip.Sundarraj <dilip.sundarraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.40.6"
python-dotenv = "^1.0.1"
streamlit = "^1.38.0"
requests = "^2.32.3"
pyttsx3 = "^2.91"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
